cleaning up api client
