+++ crates/opsml_storage/src/storage/http/client.rs
@@ -8,6 +8,7 @@ use opsml_types::contracts::FileInfo;
 use opsml_types::StorageType;
 use opsml_utils::FileUtils;
 use std::path::{Path, PathBuf};
+use std::sync::Arc;
 use tracing::debug;
 pub struct HttpFSStorageClient {
     pub client: HttpStorageClient,
@@ -21,9 +22,9 @@ impl HttpFSStorageClient {
         "HttpFSStorageClient"
     }
 
-    pub async fn new(settings: &mut OpsmlStorageSettings) -> Result<Self, StorageError> {
+    pub async fn new(api_client: Arc<OpsmlApiClient>) -> Result<Self, StorageError> {
         Ok(HttpFSStorageClient {
-            client: HttpStorageClient::new(settings).await.map_err(|e| {
+            client: HttpStorageClient::new(api_client).await.map_err(|e| {
                 StorageError::Error(format!("Failed to create http storage client {}", e))
             })?,
         })
